refactor(desktop,host-service): move all error translation into procedure bodies
Kills the parallel error-translation subsystem the first pass introduced:

- translate-git-errors.ts and the shared errno-trpc modules are gone;
  each procedure body catches domain errors and maps them to TRPCError
  inline, so a procedure's full error behavior reads in one place.
- Utils throw domain errors only (git.ts's TRPCError throws reverted to
  GitEnvironmentError), which also gets @trpc/server out of the
  git-task-worker bundle.
- runGitTask rehydrates NotGitRepoError/GitEnvironmentError from the
  worker's serialized {name, message}, so procedure bodies use instanceof
  everywhere — the name-string checks are confined to the serialization
  boundary where they belong.
- delete.ts needed nothing: its existing body-level catches already
  degrade unreadable repos to canDelete:false; create.ts gained inline
  arms around safeCheckoutBranch/getPrInfo/handleNewWorktree.
