+++ apps/desktop/src/lib/trpc/routers/changes/status.ts
@@ -1,6 +1,10 @@
+import { TRPCError } from "@trpc/server";
 import type { ChangedFile, GitChangesStatus } from "shared/changes-types";
 import { z } from "zod";
 import { publicProcedure, router } from "../..";
+import { pathExistsCached } from "../utils/path-exists-cache";
+import { NotGitRepoError } from "../workspaces/utils/git";
+import { GitEnvironmentError } from "../workspaces/utils/git-errors";
 import { assertRegisteredWorktree } from "./security/path-validation";
 import { getPersistedWorktreeBaseBranch } from "./utils/effective-base-branch";
 import {
@@ -11,10 +15,6 @@ import {
 	setCachedStatus,
 	setInFlightStatus,
 } from "./utils/status-cache";
-import {
-	assertWorktreeOnDisk,
-	translateGitTaskError,
-} from "./utils/translate-git-errors";
 import { runGitTask } from "./workers/git-task-runner";
 
 export const createStatusRouter = () => {
@@ -28,7 +28,18 @@ export const createStatusRouter = () => {
 			)
 			.query(async ({ input }): Promise<GitChangesStatus> => {
 				assertRegisteredWorktree(input.worktreePath);
-				assertWorktreeOnDisk(input.worktreePath);
+				// A worktree deleted outside the app is routine; fail fast instead of
+				// spawning a doomed git task every poll.
+				if (!pathExistsCached(input.worktreePath)) {
+					throw new TRPCError({
+						code: "NOT_FOUND",
+						message: `Worktree no longer exists on disk: ${input.worktreePath}`,
+						cause: {
+							kind: "WORKTREE_MISSING",
+							worktreePath: input.worktreePath,
+						},
+					});
+				}
 
 				const defaultBranch = input.defaultBranch || undefined;
 				const cacheKey = makeStatusCacheKey(input.worktreePath, defaultBranch);
@@ -67,7 +78,21 @@ export const createStatusRouter = () => {
 						}
 						return result;
 					} catch (error) {
-						translateGitTaskError(error);
+						if (error instanceof NotGitRepoError) {
+							throw new TRPCError({
+								code: "BAD_REQUEST",
+								message: error.message,
+								cause: { kind: "NOT_GIT_REPO" },
+							});
+						}
+						if (error instanceof GitEnvironmentError) {
+							throw new TRPCError({
+								code: "PRECONDITION_FAILED",
+								message: error.message,
+								cause: { kind: "GIT_ENVIRONMENT" },
+							});
+						}
+						throw error;
 					}
 				})();
 
@@ -105,7 +130,21 @@ export const createStatusRouter = () => {
 						},
 					);
 				} catch (error) {
-					translateGitTaskError(error);
+					if (error instanceof NotGitRepoError) {
+						throw new TRPCError({
+							code: "BAD_REQUEST",
+							message: error.message,
+							cause: { kind: "NOT_GIT_REPO" },
+						});
+					}
+					if (error instanceof GitEnvironmentError) {
+						throw new TRPCError({
+							code: "PRECONDITION_FAILED",
+							message: error.message,
+							cause: { kind: "GIT_ENVIRONMENT" },
+						});
+					}
+					throw error;
 				}
 			}),
 	});
