+++ apps/desktop/src/lib/trpc/routers/ui-state/index.ts
@@ -1,8 +1,25 @@
+import { TRPCError } from "@trpc/server";
 import { appState } from "main/lib/app-state";
 import type { TabsState, ThemeState } from "main/lib/app-state/schemas";
 import { z } from "zod";
 import { publicProcedure, router } from "../..";
-import { withFsErrorTranslation } from "../../utils/errno-trpc";
+
+// Full disks and permission walls on the user's state file are their
+// environment, not bugs.
+async function writeAppState(): Promise<void> {
+	try {
+		await appState.write();
+	} catch (error) {
+		const errno = (error as NodeJS.ErrnoException | null)?.code;
+		if (errno === "ENOSPC" || errno === "EACCES" || errno === "EPERM") {
+			throw new TRPCError({
+				code: errno === "ENOSPC" ? "PRECONDITION_FAILED" : "FORBIDDEN",
+				message: error instanceof Error ? error.message : String(error),
+			});
+		}
+		throw error;
+	}
+}
 
 /**
  * Zod schema for FileViewerState persistence.
@@ -261,7 +278,7 @@ export const createUiStateRouter = () => {
 				.input(tabsStateSchema)
 				.mutation(async ({ input }) => {
 					appState.data.tabsState = input;
-					await withFsErrorTranslation(() => appState.write());
+					await writeAppState();
 					return { success: true };
 				}),
 		}),
@@ -276,7 +293,7 @@ export const createUiStateRouter = () => {
 				.input(themeStateSchema)
 				.mutation(async ({ input }) => {
 					appState.data.themeState = input;
-					await withFsErrorTranslation(() => appState.write());
+					await writeAppState();
 					return { success: true };
 				}),
 		}),
