+++ apps/desktop/src/lib/trpc/routers/workspaces/utils/git.ts
@@ -10,10 +10,8 @@ import {
 	sanitizeBranchName,
 	sanitizeBranchNameWithMaxLength,
 } from "@superset/shared/workspace-launch";
-import { TRPCError } from "@trpc/server";
 import friendlyWords from "friendly-words";
 import type { StatusResult } from "simple-git";
-import { translateGitTaskError } from "../../changes/utils/translate-git-errors";
 import { runWithPostCheckoutHookTolerance } from "../../utils/git-hook-tolerance";
 import { execGitWithShellPath, getSimpleGitWithShellPath } from "./git-client";
 import { GitEnvironmentError } from "./git-errors";
@@ -23,8 +21,8 @@ import { resolveTrackingRemoteName } from "./upstream-ref";
 const execFileAsync = promisify(execFile);
 
 export class NotGitRepoError extends Error {
-	constructor(repoPath: string) {
-		super(`Not a git repository: ${repoPath}`);
+	constructor(message: string) {
+		super(message);
 		this.name = "NotGitRepoError";
 	}
 }
@@ -231,7 +229,7 @@ export async function getStatusNoLock(repoPath: string): Promise<StatusResult> {
 			}
 			const stderr = error.stderr || error.message || "";
 			if (stderr.includes("not a git repository")) {
-				throw new NotGitRepoError(repoPath);
+				throw new NotGitRepoError(`Not a git repository: ${repoPath}`);
 			}
 		}
 		// maxBuffer overflows, timeouts, permission walls: pathological working
@@ -637,14 +635,12 @@ export async function createWorktree(
 			console.error(
 				`Git lock file error during worktree creation: ${errorMessage}`,
 			);
-			throw new TRPCError({
-				code: "PRECONDITION_FAILED",
-				message:
-					`Failed to create worktree: The git repository is locked by another process. ` +
+			throw new GitEnvironmentError(
+				`Failed to create worktree: The git repository is locked by another process. ` +
 					`This usually happens when another git operation is in progress, or a previous operation crashed. ` +
 					`Please wait for the other operation to complete, or manually remove the lock file ` +
 					`(e.g., .git/config.lock or .git/index.lock) if you're sure no git operations are running.`,
-			});
+			);
 		}
 
 		console.error(`Failed to create worktree: ${errorMessage}`);
@@ -728,14 +724,12 @@ export async function createWorktreeFromExistingBranch({
 			console.error(
 				`Git lock file error during worktree creation: ${errorMessage}`,
 			);
-			throw new TRPCError({
-				code: "PRECONDITION_FAILED",
-				message:
-					`Failed to create worktree: The git repository is locked by another process. ` +
+			throw new GitEnvironmentError(
+				`Failed to create worktree: The git repository is locked by another process. ` +
 					`This usually happens when another git operation is in progress, or a previous operation crashed. ` +
 					`Please wait for the other operation to complete, or manually remove the lock file ` +
 					`(e.g., .git/config.lock or .git/index.lock) if you're sure no git operations are running.`,
-			});
+			);
 		}
 
 		// Check if the branch is already checked out in another worktree
@@ -743,12 +737,10 @@ export async function createWorktreeFromExistingBranch({
 			lowerError.includes("already checked out") ||
 			lowerError.includes("is already used by worktree")
 		) {
-			throw new TRPCError({
-				code: "PRECONDITION_FAILED",
-				message:
-					`Branch "${branch}" is already checked out in another worktree. ` +
+			throw new GitEnvironmentError(
+				`Branch "${branch}" is already checked out in another worktree. ` +
 					`Each branch can only be checked out in one worktree at a time.`,
-			});
+			);
 		}
 
 		console.error(`Failed to create worktree: ${errorMessage}`);
@@ -842,7 +834,7 @@ export async function getGitRoot(path: string): Promise<string> {
 	} catch (error) {
 		const message = error instanceof Error ? error.message : String(error);
 		if (message.toLowerCase().includes("not a git repository")) {
-			throw new NotGitRepoError(path);
+			throw new NotGitRepoError(`Not a git repository: ${path}`);
 		}
 		throw error;
 	}
@@ -1128,12 +1120,8 @@ export async function getAheadBehindCount({
 export async function hasUncommittedChanges(
 	worktreePath: string,
 ): Promise<boolean> {
-	try {
-		const status = await getStatusNoLock(worktreePath);
-		return !status.isClean();
-	} catch (error) {
-		translateGitTaskError(error);
-	}
+	const status = await getStatusNoLock(worktreePath);
+	return !status.isClean();
 }
 
 export async function hasUnpushedCommits(
@@ -1651,10 +1639,9 @@ export async function safeCheckoutBranch(
 
 	const safety = await checkBranchCheckoutSafety(repoPath);
 	if (!safety.safe) {
-		throw new TRPCError({
-			code: "PRECONDITION_FAILED",
-			message: safety.error,
-		});
+		throw new GitEnvironmentError(
+			safety.error ?? "Branch checkout is not safe",
+		);
 	}
 
 	await checkoutBranch(repoPath, branch);
@@ -1771,28 +1758,23 @@ export async function getPrInfo({
 	} catch (error) {
 		if (isExecFileException(error)) {
 			if (error.code === "ENOENT") {
-				throw new TRPCError({
-					code: "PRECONDITION_FAILED",
-					message:
-						"GitHub CLI (gh) is not installed. Please install it from https://cli.github.com/",
-				});
+				throw new GitEnvironmentError(
+					"GitHub CLI (gh) is not installed. Please install it from https://cli.github.com/",
+				);
 			}
 			const stderr = error.stderr || error.message || "";
 			if (stderr.includes("not logged in")) {
-				throw new TRPCError({
-					code: "PRECONDITION_FAILED",
-					message:
-						"Not logged in to GitHub CLI. Please run 'gh auth login' first.",
-				});
+				throw new GitEnvironmentError(
+					"Not logged in to GitHub CLI. Please run 'gh auth login' first.",
+				);
 			}
 			if (
 				stderr.includes("Could not resolve") ||
 				stderr.includes("not found")
 			) {
-				throw new TRPCError({
-					code: "NOT_FOUND",
-					message: `PR #${prNumber} not found in ${owner}/${repo}`,
-				});
+				throw new GitEnvironmentError(
+					`PR #${prNumber} not found in ${owner}/${repo}`,
+				);
 			}
 		}
 		throw new Error(
@@ -1903,10 +1885,9 @@ export async function createWorktreeFromPr({
 			lowerError.includes("already checked out") ||
 			lowerError.includes("is already used by worktree")
 		) {
-			throw new TRPCError({
-				code: "PRECONDITION_FAILED",
-				message: `This PR's branch is already checked out in another worktree.`,
-			});
+			throw new GitEnvironmentError(
+				`This PR's branch is already checked out in another worktree.`,
+			);
 		}
 		throw new Error(`Failed to create worktree from PR: ${errorMessage}`);
 	}
